Functionalities of ExperimentManager class merged with FederatedRuntime.
