+++ openfl/experimental/interface/interactive_api/__init__.py
@@ -3,4 +3,3 @@
 """openfl.experimental.interface.interactive_api package."""
 
 from openfl.experimental.interface.interactive_api.federation import Federation
-from openfl.experimental.interface.interactive_api.experiment_manager import ExperimentManager, ExperimentStatus
