+++ openfl/experimental/runtime/__init__.py
@@ -4,6 +4,6 @@
 
 """ openfl.experimental.runtime package Runtime class."""
 
-from openfl.experimental.runtime.federated_runtime import FederatedRuntime
+from openfl.experimental.runtime.federated_runtime import ExperimentStatus, FederatedRuntime
 from openfl.experimental.runtime.local_runtime import LocalRuntime
 from openfl.experimental.runtime.runtime import Runtime
